# Modularize TRUELENGTH saving/reallocating helpers