--- a/src/order-truelength.c
+++ b/src/order-truelength.c
@@ -29,23 +29,28 @@
  *
  * Pair with `PROTECT_TRUELENGTH_INFO()` in the caller
  */
-struct truelength_info* new_truelength_info(r_ssize max_size_alloc) {
+struct truelength_info* new_truelength_info(r_ssize n_max) {
   SEXP self = PROTECT(r_new_raw(sizeof(struct truelength_info)));
   struct truelength_info* p_truelength_info = (struct truelength_info*) RAW(self);
 
   p_truelength_info->self = self;
 
   p_truelength_info->strings = vctrs_shared_empty_chr;
-  p_truelength_info->lengths = vctrs_shared_empty_raw;
+  p_truelength_info->truelengths = vctrs_shared_empty_raw;
+  p_truelength_info->n_strings_alloc = 0;
+  p_truelength_info->n_strings_used = 0;
+
   p_truelength_info->uniques = vctrs_shared_empty_chr;
+  p_truelength_info->n_uniques_alloc = 0;
+  p_truelength_info->n_uniques_used = 0;
+
   p_truelength_info->sizes = vctrs_shared_empty_int;
   p_truelength_info->sizes_aux = vctrs_shared_empty_int;
-
-  p_truelength_info->size_alloc = 0;
-  p_truelength_info->max_size_alloc = max_size_alloc;
-  p_truelength_info->size_used = 0;
-
+  p_truelength_info->n_sizes_alloc = 0;
+  p_truelength_info->n_sizes_used = 0;
   p_truelength_info->max_string_size = 0;
+
+  p_truelength_info->n_max = n_max;
 
   UNPROTECT(1);
   return p_truelength_info;
@@ -54,112 +59,79 @@ struct truelength_info* new_truelength_info(r_ssize max_size_alloc) {
 // -----------------------------------------------------------------------------
 
 /*
- * Reset the truelengths of all unique strings captured in `strings` using
- * the original truelengths in `lengths`.
+ * First, reset the TRUELENGTHs of all unique CHARSXPs in `uniques` to 0, which
+ * is the default used by R.
  *
- * Will be called after each character data frame column is processed, and
+ * Then, reset the TRUELENGTHs of all CHARSXPs in `strings` to their original
+ * value contained in `truelengths`. There should be very few of these,
+ * if any, as R doesn't typically use the TRUELENGTH slot of CHARSXPs. One
+ * exception seems to be very simple strings, such as `"a"`, which R probably
+ * adds to the cache at startup, and sets their TRUELENGTH value for some
+ * reason.
+ *
+ * It is important to reset `uniques` first, then `strings`, as the CHARSXPs
+ * in `strings` are, by definition, also in `uniques`.
+ *
+ * This will be called after each character data frame column is processed, and
  * at the end of `chr_order()` for a single character vector.
  */
 void truelength_reset(struct truelength_info* p_truelength_info) {
-  r_ssize size = p_truelength_info->size_used;
+  r_ssize n_uniques_used = p_truelength_info->n_uniques_used;
+  r_ssize n_strings_used = p_truelength_info->n_strings_used;
 
-  for (r_ssize i = 0; i < size; ++i) {
+  // First reset uniques
+  for (r_ssize i = 0; i < n_uniques_used; ++i) {
+    SEXP unique = p_truelength_info->p_uniques[i];
+    SET_TRUELENGTH(unique, 0);
+  }
+
+  // Then reset strings
+  for (r_ssize i = 0; i < n_strings_used; ++i) {
     SEXP string = p_truelength_info->p_strings[i];
-    r_ssize length = p_truelength_info->p_lengths[i];
-
-    SET_TRUELENGTH(string, length);
+    r_ssize truelength = p_truelength_info->p_truelengths[i];
+    SET_TRUELENGTH(string, truelength);
   }
 
   // Also reset vector specific details
-  p_truelength_info->size_used = 0;
+  p_truelength_info->n_uniques_used = 0;
+  p_truelength_info->n_strings_used = 0;
+  p_truelength_info->n_sizes_used = 0;
   p_truelength_info->max_string_size = 0;
 }
 
 // -----------------------------------------------------------------------------
 
-static void truelength_realloc(struct truelength_info* p_truelength_info);
+static r_ssize truelength_realloc_size(r_ssize n_x, r_ssize n_max);
 
-/*
- * Saves a unique CHARSXP `x` along with its original truelength and
- * its "size" (i.e the number of characters). Will be reset later with
- * `truelength_reset()`.
- */
-void truelength_save(SEXP x,
-                     r_ssize truelength,
-                     r_ssize size,
-                     struct truelength_info* p_truelength_info) {
-  // Reallocate as needed
-  if (p_truelength_info->size_used == p_truelength_info->size_alloc) {
-    truelength_realloc(p_truelength_info);
-  }
+static inline SEXP truelengths_resize(SEXP x, r_ssize x_size, r_ssize size);
 
-  p_truelength_info->p_strings[p_truelength_info->size_used] = x;
-  p_truelength_info->p_lengths[p_truelength_info->size_used] = truelength;
-  p_truelength_info->p_uniques[p_truelength_info->size_used] = x;
-  p_truelength_info->p_sizes[p_truelength_info->size_used] = size;
-
-  ++p_truelength_info->size_used;
-}
-
-// -----------------------------------------------------------------------------
-
-static r_ssize truelength_realloc_size(struct truelength_info* p_truelength_info);
-
-static inline SEXP lengths_resize(SEXP x, r_ssize x_size, r_ssize size);
-
-/*
- * Extend the vectors in `truelength_info`.
- * Reprotects itself.
- */
-static
-void truelength_realloc(struct truelength_info* p_truelength_info) {
-  r_ssize size = truelength_realloc_size(p_truelength_info);
+void truelength_realloc_strings(struct truelength_info* p_truelength_info) {
+  r_ssize size = truelength_realloc_size(
+    p_truelength_info->n_strings_alloc,
+    p_truelength_info->n_max
+  );
 
   p_truelength_info->strings = chr_resize(
     p_truelength_info->strings,
-    p_truelength_info->size_used,
+    p_truelength_info->n_strings_alloc,
     size
   );
   REPROTECT(p_truelength_info->strings, p_truelength_info->strings_pi);
   p_truelength_info->p_strings = STRING_PTR(p_truelength_info->strings);
 
-  p_truelength_info->lengths = lengths_resize(
-    p_truelength_info->lengths,
-    p_truelength_info->size_used,
+  p_truelength_info->truelengths = truelengths_resize(
+    p_truelength_info->truelengths,
+    p_truelength_info->n_strings_alloc,
     size
   );
-  REPROTECT(p_truelength_info->lengths, p_truelength_info->lengths_pi);
-  p_truelength_info->p_lengths = (r_ssize*) RAW(p_truelength_info->lengths);
+  REPROTECT(p_truelength_info->truelengths, p_truelength_info->truelengths_pi);
+  p_truelength_info->p_truelengths = (r_ssize*) RAW(p_truelength_info->truelengths);
 
-  p_truelength_info->uniques = chr_resize(
-    p_truelength_info->uniques,
-    p_truelength_info->size_used,
-    size
-  );
-  REPROTECT(p_truelength_info->uniques, p_truelength_info->uniques_pi);
-  p_truelength_info->p_uniques = STRING_PTR(p_truelength_info->uniques);
-
-  p_truelength_info->sizes = int_resize(
-    p_truelength_info->sizes,
-    p_truelength_info->size_used,
-    size
-  );
-  REPROTECT(p_truelength_info->sizes, p_truelength_info->sizes_pi);
-  p_truelength_info->p_sizes = INTEGER(p_truelength_info->sizes);
-
-  p_truelength_info->sizes_aux = int_resize(
-    p_truelength_info->sizes_aux,
-    p_truelength_info->size_used,
-    size
-  );
-  REPROTECT(p_truelength_info->sizes_aux, p_truelength_info->sizes_aux_pi);
-  p_truelength_info->p_sizes_aux = INTEGER(p_truelength_info->sizes_aux);
-
-  p_truelength_info->size_alloc = size;
+  p_truelength_info->n_strings_alloc = size;
 }
 
 static inline
-SEXP lengths_resize(SEXP x, r_ssize x_size, r_ssize size) {
+SEXP truelengths_resize(SEXP x, r_ssize x_size, r_ssize size) {
   return raw_resize(
     x,
     x_size * sizeof(r_ssize),
@@ -169,28 +141,71 @@ SEXP lengths_resize(SEXP x, r_ssize x_size, r_ssize size) {
 
 // -----------------------------------------------------------------------------
 
-static
-r_ssize truelength_realloc_size(struct truelength_info* p_truelength_info) {
-  r_ssize size_alloc = p_truelength_info->size_alloc;
-  r_ssize max_size_alloc = p_truelength_info->max_size_alloc;
+void truelength_realloc_uniques(struct truelength_info* p_truelength_info) {
+  r_ssize size = truelength_realloc_size(
+    p_truelength_info->n_uniques_alloc,
+    p_truelength_info->n_max
+  );
 
+  p_truelength_info->uniques = chr_resize(
+    p_truelength_info->uniques,
+    p_truelength_info->n_uniques_alloc,
+    size
+  );
+  REPROTECT(p_truelength_info->uniques, p_truelength_info->uniques_pi);
+  p_truelength_info->p_uniques = STRING_PTR(p_truelength_info->uniques);
+
+  p_truelength_info->n_uniques_alloc = size;
+}
+
+// -----------------------------------------------------------------------------
+
+void truelength_realloc_sizes(struct truelength_info* p_truelength_info) {
+  r_ssize size = truelength_realloc_size(
+    p_truelength_info->n_sizes_alloc,
+    p_truelength_info->n_max
+  );
+
+  p_truelength_info->sizes = int_resize(
+    p_truelength_info->sizes,
+    p_truelength_info->n_sizes_alloc,
+    size
+  );
+  REPROTECT(p_truelength_info->sizes, p_truelength_info->sizes_pi);
+  p_truelength_info->p_sizes = INTEGER(p_truelength_info->sizes);
+
+  p_truelength_info->sizes_aux = int_resize(
+    p_truelength_info->sizes_aux,
+    p_truelength_info->n_sizes_alloc,
+    size
+  );
+  REPROTECT(p_truelength_info->sizes_aux, p_truelength_info->sizes_aux_pi);
+  p_truelength_info->p_sizes_aux = INTEGER(p_truelength_info->sizes_aux);
+
+  p_truelength_info->n_sizes_alloc = size;
+}
+
+// -----------------------------------------------------------------------------
+
+static
+r_ssize truelength_realloc_size(r_ssize n_x, r_ssize n_max) {
   // First allocation
-  if (size_alloc == 0) {
-    if (TRUELENGTH_SIZE_ALLOC_DEFAULT < max_size_alloc) {
+  if (n_x == 0) {
+    if (TRUELENGTH_SIZE_ALLOC_DEFAULT < n_max) {
       return TRUELENGTH_SIZE_ALLOC_DEFAULT;
     } else {
-      return max_size_alloc;
+      return n_max;
     }
   }
 
   // Avoid potential overflow when doubling size
-  uint64_t new_size_alloc = ((uint64_t) size_alloc) * 2;
+  uint64_t n_new = ((uint64_t) n_x) * 2;
 
   // Clamp maximum allocation size to the size of the input
-  if (new_size_alloc > max_size_alloc) {
-    return max_size_alloc;
+  if (n_new > n_max) {
+    return n_max;
   }
 
   // Can now safely cast back to `r_ssize`
-  return (r_ssize) new_size_alloc;
+  return (r_ssize) n_new;
 }