For `vec_order_radix()`'s `struct truelength_info`, we only have to store the original TRUELENGTH value in `strings` / `truelengths` if R is actually using it, i.e. the value is >0, otherwise we can always reset it to the default value of 0. It is rare for R to use the TRUELENGTH, so we should rarely have to reallocate the `strings` and `truelengths` vectors. This saves us some time with large vectors when most of the strings are unique, because we previously saved every unique value and its TRUELENGTH.

Base R actually does this optimization too, I just didn't see the benefit of it until now.

I've also separated out the saving/reallocating of individual CHARSXP `sizes` into its own helper as well. I've got an upcoming PR that allows for ordering character vectors by appearance, which doesn't require the `sizes` vector, so this is preparation for that. That PR will be used for reimplementing the dictionary functions.

In the numbers below, I'm mainly interested in the big drop in memory usage required to sort a character vector with lots of unique values

``` r
set.seed(123)

vec_order_radix <- vctrs:::vec_order_radix

random_strings <- function(size, min_length, max_length) {
  lengths <- rlang::seq2(min_length, max_length)
  stringi::stri_rand_strings(
    size,
    sample(lengths, size = size, replace = TRUE)
  )
}

# all unique strings!
x <- random_strings(size = 1e7, min_length = 5, max_length = 20)

# before
bench::mark(vec_order_radix(x), iterations = 20)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_order_radix(x)    5.86s    5.97s     0.162     856MB    0.341

# after
bench::mark(vec_order_radix(x), iterations = 20)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression              min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>         <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_order_radix(x)    5.12s    5.23s     0.185     547MB    0.222
```